+++ packages/vm/src/compatibility.rs
@@ -22,6 +22,8 @@ const SUPPORTED_IMPORTS: &[&str] = &[
     "env.addr_humanize",
     "env.secp256k1_verify",
     "env.secp256k1_recover_pubkey",
+    "env.secp256r1_verify",
+    "env.secp256r1_recover_pubkey",
     "env.ed25519_verify",
     "env.ed25519_batch_verify",
     "env.debug",
@@ -635,6 +637,8 @@ mod tests {
             (import "env" "addr_humanize" (func (param i32 i32) (result i32)))
             (import "env" "secp256k1_verify" (func (param i32 i32 i32) (result i32)))
             (import "env" "secp256k1_recover_pubkey" (func (param i32 i32 i32) (result i64)))
+            (import "env" "secp256r1_verify" (func (param i32 i32 i32) (result i32)))
+            (import "env" "secp256r1_recover_pubkey" (func (param i32 i32 i32) (result i64)))
             (import "env" "ed25519_verify" (func (param i32 i32 i32) (result i32)))
             (import "env" "ed25519_batch_verify" (func (param i32 i32 i32) (result i32)))
         )"#,
@@ -654,6 +658,8 @@ mod tests {
             (import "env" "addr_humanize" (func (param i32 i32) (result i32)))
             (import "env" "secp256k1_verify" (func (param i32 i32 i32) (result i32)))
             (import "env" "secp256k1_recover_pubkey" (func (param i32 i32 i32) (result i64)))
+            (import "env" "secp256r1_verify" (func (param i32 i32 i32) (result i32)))
+            (import "env" "secp256r1_recover_pubkey" (func (param i32 i32 i32) (result i64)))
             (import "env" "ed25519_verify" (func (param i32 i32 i32) (result i32)))
             (import "env" "ed25519_batch_verify" (func (param i32 i32 i32) (result i32)))
             (import "env" "spam01" (func (param i32 i32) (result i32)))
@@ -746,8 +752,6 @@ mod tests {
             (import "env" "spam88" (func (param i32 i32) (result i32)))
             (import "env" "spam89" (func (param i32 i32) (result i32)))
             (import "env" "spam90" (func (param i32 i32) (result i32)))
-            (import "env" "spam91" (func (param i32 i32) (result i32)))
-            (import "env" "spam92" (func (param i32 i32) (result i32)))
         )"#,
         )
         .unwrap();