+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- cosmwasm-vm: Add `secp256r1_verify` and `secp256r1_recover_pubkey` imports for
+  ECDSA signature verification over secp256r1. ([#1983])
+
+[#1983]: https://github.com/CosmWasm/cosmwasm/pull/1983
+
 ### Changed
 
 - cosmwasm-std: Enable `add_event` and `add_events` functions to process types
@@ -13,7 +20,7 @@ and this project adheres to
 
 [#2044]: https://github.com/CosmWasm/cosmwasm/pull/2044
 
-## [2.0.0-rc.1] - 2023-02-09
+## [2.0.0] - 2024-03-12
 
 ### Fixed
 
@@ -760,8 +767,8 @@ and this project adheres to
 The CHANGELOG for versions before 1.0.0 was moved to
 [CHANGELOG-pre1.0.0.md](./CHANGELOG-pre1.0.0.md).
 
-[unreleased]: https://github.com/CosmWasm/cosmwasm/compare/v2.0.0-rc.1...HEAD
-[2.0.0-rc.1]: https://github.com/CosmWasm/cosmwasm/compare/v1.5.0...v2.0.0-rc.1
+[unreleased]: https://github.com/CosmWasm/cosmwasm/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/CosmWasm/cosmwasm/compare/v1.5.0...v2.0.0
 [1.5.0]: https://github.com/CosmWasm/cosmwasm/compare/v1.4.1...v1.5.0
 [1.4.1]: https://github.com/CosmWasm/cosmwasm/compare/v1.4.0...v1.4.1
 [1.4.0]: https://github.com/CosmWasm/cosmwasm/compare/v1.3.3...v1.4.0